Remove the `paused` constructor argument for `Pipe`.
It was pretty pointless. The two constructors that take a `paused`
argument do so because it's sensible for them to start out paused, but
it seemed odd to mandate it.

Originally, I thought all the constructors in the module would have
`paused` arguments, but that idea fell by the wayside.

This change now makes things consistent in that, if there *is* a
`paused` argument, the default is always `true`.