Start requiring `typ`.
There are already a few cases where its assertions would be handy, and
I've just avoided adding it as a dependency because it seemed like
there weren't enough cases to justify its load. But with the current
development arc, I think it's tipped over the edge and is now worth
including.