Change the default for `paused` on `Cat` and `Valve`.
These had been paused by default, but that can be pretty confusing.

It *used* to be that there was no compelling reason not to start out
paused, but at this point, the module does event sanitization, and that
means there is now a use case for paused == `false` that can be
expected to be common enough.

With this change, the only class that starts out paused is `Blip`,
because there is truly no sense in doing otherwise.